requirements.txt
torchvision
jupyterlab
pandas
numpy
sentencepiece
tiktoken
tqdm